requirements.txt
# requirements.txt

anyio==4.1.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==3.1.0
comm==0.2.0
cryptography==43.0.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.1
django-allauth==0.57.0
django-ckeditor==6.7.1
django-js-asset==2.1.0
django-ckeditor-5
executing==2.0.1
fastjsonschema==2.19.0
fqdn==1.5.1
idna==3.7
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mysqlclient
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.2.2
notebook_shim==0.2.3
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.3
pexpect==4.9.0
Pillow
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.7.0
PyMySQL==1.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python3-openid==3.2.0
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
requests==2.32.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
Send2Trash==1.8.2
setuptools==72.2.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4.1
traitlets==5.14.0
types-python-dateutil==2.8.19.14
tzdata==2023.3
uri-template==1.3.0
urllib3==2.2.2
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
aiohttp
pyyaml
ollama
pypdfium2
pytesseract
pdf2image
poppler
pyocr
opencv-python
numpy
python-dotenv